pyproject.toml
[project]
name = "serp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.179.0",
    "python-dotenv>=1.1.1",
    "yt-dlp>=2025.8.20",
]
